requirements.txt
msal
requests
pdfplumber
pytesseract
pillow
pdf2image
openpyxl
python-dotenv
